GH Issue #2: Raven flash-drives
This took quite a bit longer than I thought it would take but I think
it's worth it.

The previous major commit added raven-configure; a helper script for the
raven template flake.

This commit adds working disko profiles and expands the raven-configure
script to handle selecting one of them, or installing Raven on top of an
existing NixOS system.

The raven-iso works and boots, and comes included with a basic
shell-script bin for cloning the Raven repository and executing the
raven-configure script.

While the installation itself does not proceed automatically,
instructions are printed at the end of the raven-configure script
depending on what disk profile was selected. I was not yet ready to
commit to an automated installation procedure. That may change in the
future.

What is important now is Raven has an ISO, and has the ability to install
itself in one-shot from that ISO without regard for the NixOS ISO, and
has a guided interactive script that doesn't suck. Raven is, at this
point, a great Nix flake for those looking to get into Nix flakes, and a
good way to install a new Linux OS in one shot.

Build the ISO with `nix build ~/repos/raven-machines#nixosConfigurations.raven-iso.config.system.build.isoImage`