Added raven-configure script
This commit moves the check script to raven-configure. This is an
interactive script that can be run non-interactively if the proper
environment variables are set.

This makes editing a raven template a bit easier.

This also changes the format of some of the "REPLACEME" text around the
nix files in this repo.

TODO: Drive settings