BREAKING: Attachment architecture
raven-home and raven-gaming are no more. raven-desktop is back and
raven-gaming is reduced to an attachment for raven-desktop

The change after this will be documentation updates to indicate the new
pattern, but it's basically...

patterns:       raven-minimal -> raven-terminal -> raven-desktop
attachments:                                       └── gaming

Attachments, for now, are expected to live under compliant patterns in an
"attachments" sub-directory. Same general organization; system stuff at
the main level and home-manager configuration in a home-manager
subdirectory.

TODO: Integrate with raven-configure or some other convenient way of
selecting attachments. A user shouldn't have to dig deep into patterns
of Raven to add "gaming" or "production" to their desktop suite.

Hyprland would provide a dope way to select attachments graphically...

But for now the only attachment is gaming, and it's... well not
lightweight, but certainly not onerous on a desktop system. It's
configured to be defaultly attached to raven-desktop for now.

It's fine!!!